+++ packages/router/test/globalTestHook.ts
@@ -16,7 +16,10 @@ import { Web3Signer } from "../src/adapters/web3signer";
 import { ContractReader } from "../src/adapters/subgraph";
 import { ContractWriter } from "../src/adapters/contract";
 import * as RouterFns from "../src/router";
+import { AuctionCache } from "../src/adapters/cache/auction";
+import { RouterCache } from "../src/adapters/cache";
 
+export let cacheMock: RouterCache;
 export let txServiceMock: SinonStubbedInstance<TransactionService>;
 export let messagingMock: SinonStubbedInstance<RouterNxtpNatsMessagingService>;
 export let contractReaderMock: ContractReader;
@@ -72,6 +75,13 @@ export const mochaHooks = {
       ]),
     };
 
+    const auctionsCacheMock = createStubInstance(AuctionCache);
+    auctionsCacheMock.getOutstandingLiquidity.returns(BigNumber.from("0"));
+    auctionsCacheMock.addBid.resolves();
+    cacheMock = {
+      auctions: auctionsCacheMock as any,
+    };
+
     contractWriterMock = {
       prepareRouterContract: stub().resolves(txReceiptMock),
       prepareTransactionManager: stub().resolves(txReceiptMock),
@@ -95,6 +105,7 @@ export const mochaHooks = {
       isRouterContract: undefined,
       routerAddress,
       signerAddress,
+      cache: cacheMock,
       chainAssetSwapPoolMap: chainAssetSwapPoolMapMock,
     };